Split up Nimbus.kt into smaller files
